dj - sped up tests by rearranging spring annotations
