+++ src/main/java/edu/ucsb/cs156/happiercows/HappierCowsApplication.java
@@ -20,49 +20,13 @@ import edu.ucsb.cs156.happiercows.services.wiremock.WiremockService;
 import lombok.extern.slf4j.Slf4j;
 
 @SpringBootApplication
-@EnableJpaAuditing(dateTimeProviderRef = "utcDateTimeProvider")
 @EnableAsync
 @EnableScheduling
 @Slf4j
 public class HappierCowsApplication {
-
-  @Autowired
-  WiremockService wiremockService;
-
-  /**
-   * When using the wiremock profile, this method will call the code needed to set up the wiremock services
-   */
-  @Profile("wiremock")
-  @Bean
-  public ApplicationRunner wiremockApplicationRunner() {
-    return arg -> {
-      log.info("wiremock mode");
-      wiremockService.init();
-      log.info("wiremockApplicationRunner completed");
-    };
-  }
-
-  /**
-   * Hook that can be used to set up any services needed for development
-   */
-  @Profile("development")
-  @Bean
-  public ApplicationRunner developmentApplicationRunner() {
-    return arg -> {
-      log.info("development mode");
-      log.info("developmentApplicationRunner completed");
-    };
-  }
   
   public static void main(String[] args) {
     SpringApplication.run(HappierCowsApplication.class, args);
-  }
-  @Bean
-  public DateTimeProvider utcDateTimeProvider() {
-      return () -> {
-        ZonedDateTime now = ZonedDateTime.now();
-        return Optional.of(now);
-      };
   }
 
   // See: https://www.baeldung.com/spring-security-async-principal-propagation
