+++ src/main/java/edu/ucsb/cs156/happiercows/interceptors/RoleUserInterceptorAppConfig.java
@@ -1,11 +1,12 @@
 package edu.ucsb.cs156.happiercows.interceptors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-@Component
+@Configuration
 public class RoleUserInterceptorAppConfig implements WebMvcConfigurer {
    @Autowired
    RoleUserInterceptor roleUserInterceptor;
