+++ src/test/java/edu/ucsb/cs156/happiercows/ControllerTestCase.java
@@ -1,8 +1,13 @@
 package edu.ucsb.cs156.happiercows;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
+import edu.ucsb.cs156.happiercows.interceptors.RoleUserInterceptor;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
