+++ src/test/java/edu/ucsb/cs156/happiercows/controllers/ChatMessageControllerTests.java
@@ -8,6 +8,8 @@ import static org.mockito.Mockito.when;
 
 import static org.mockito.ArgumentMatchers.any;
 
+import edu.ucsb.cs156.happiercows.repositories.UserRepository;
+import edu.ucsb.cs156.happiercows.testconfig.TestConfig;
 import java.util.Arrays;
 import java.util.Optional;
 
@@ -43,8 +45,6 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @WebMvcTest(controllers = ChatMessageController.class)
-@Import(ChatMessageController.class)
-@AutoConfigureDataJpa
 public class ChatMessageControllerTests extends ControllerTestCase {
     
     @MockBean
@@ -52,6 +52,9 @@ public class ChatMessageControllerTests extends ControllerTestCase {
 
     @MockBean
     UserCommonsRepository userCommonsRepository;
+
+    @MockBean
+    UserRepository userRepository;
 
     @Autowired
     ObjectMapper mapper;
