+++ src/test/java/edu/ucsb/cs156/happiercows/controllers/ProfitsControllerTests.java
@@ -11,6 +11,7 @@ import edu.ucsb.cs156.happiercows.repositories.CommonsRepository;
 import edu.ucsb.cs156.happiercows.repositories.ProfitRepository;
 import edu.ucsb.cs156.happiercows.repositories.UserCommonsRepository;
 import edu.ucsb.cs156.happiercows.repositories.UserRepository;
+import edu.ucsb.cs156.happiercows.testconfig.TestConfig;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -40,8 +41,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 @WebMvcTest(controllers = ProfitsController.class)
-@Import(ProfitsController.class)
-@AutoConfigureDataJpa
 public class ProfitsControllerTests extends ControllerTestCase {
     @Autowired
     private ObjectMapper objectMapper;
