+++ src/test/java/edu/ucsb/cs156/happiercows/controllers/ReportsControllerTests.java
@@ -15,6 +15,7 @@ import edu.ucsb.cs156.happiercows.repositories.UserCommonsRepository;
 import edu.ucsb.cs156.happiercows.repositories.UserRepository;
 import edu.ucsb.cs156.happiercows.strategies.CowHealthUpdateStrategies;
 
+import edu.ucsb.cs156.happiercows.testconfig.TestConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
@@ -37,8 +38,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(controllers = ReportsController.class)
-@Import(ReportsController.class)
-@AutoConfigureDataJpa
 public class ReportsControllerTests extends ControllerTestCase {
         @Autowired
         private ObjectMapper objectMapper;
