+++ src/test/java/edu/ucsb/cs156/happiercows/controllers/UserCommonsControllerTests.java
@@ -27,7 +27,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(controllers = UserCommonsController.class)
-@AutoConfigureDataJpa
 public class UserCommonsControllerTests extends ControllerTestCase {
 
     @MockBean
